Letter now displays in box of correct size
Originally I was just placing letters into a bitmap that was hopefully big enough. This was a massive waste of space and efficiency. As well as potentially breaking with large letters.

Now the letters are given exactly the right amount of space to be drawn in.

Works with individual letters and words/sentences.